refactor: move fetchSecureToken to util
